old-version-master/channel_power/youtube/membership/requirements.txt
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
exceptiongroup==1.1.0
google-auth==2.16.1
google-auth-oauthlib==1.0.0
greenlet==2.0.1
gspread==5.7.2
idna==3.4
iniconfig==2.0.0
oauthlib==3.2.2
onetimepass==1.0.1
packaging==23.0
playwright==1.31.1
pluggy==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==9.0.4
pytest==7.2.1
pytest-base-url==2.0.0
pytest-playwright==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
